requirements.txt
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
freetype-py==2.4.0
fsspec==2024.6.1
grpcio==1.65.5
hydra-core==1.3.2
idna==3.7
imageio==2.35.1
importlib_metadata==8.2.0
ipython==8.26.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.17.0
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==5.27.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pyglet==2.0.17
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
pyrfuniverse==0.30.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
timeout-decorator==0.5.0
torch==2.4.0
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.4.6
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urdfpy==0.0.22
urllib3==2.2.2
vhacdx==0.0.8.post1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
zipp==3.20.0
